split client and server source
