+++ v.zyablov1/task32/server_source/client_handler.h
@@ -7,6 +7,7 @@
 void process_text(char *buffer, ssize_t length);
 void handle_client(int client_sock, int epoll_fd);
 int check_for_keyword(const char *buffer, const char *keyword);
+void accept_new_client(int server_sock, int epoll_fd);
 
 #endif // CLIENT_HANDLER_H
