+++ v.zyablov1/task32/server_source/server.h
@@ -8,14 +8,11 @@
 
 // Структура для хранения состояния клиента
 typedef struct {
-    int fd;
-    int active;
+	int fd;
+	int active;
 } client_state_t;
 
-int create_server_socket(const char *socket_path);
-void accept_new_client(int server_sock, int epoll_fd);
 void server_main_loop(int server_sock, int epoll_fd);
-int setup_epoll(int server_sock);
 
 #endif // SERVER_H
