+++ v.zyablov1/task32/server_source/setup_server.h
@@ -1,0 +1,5 @@
+// Создание серверного сокета
+int create_server_socket(const char *socket_path);
+
+// Создание epoll и добавление серверного сокета
+int setup_epoll(int server_sock);
